♻️ Refactor: Pascal로 작성된 컨벤션 지키지 않은 폴더명 수정 (temp)
