--- src/common/components/LogoTest.tsx
@@ -1,117 +1,0 @@
-import React from 'react';
-
-import AiManager from '@svgs/component-set/AiManagerIcon.svg';
-import AiQuestionHover from '@svgs/component-set/AiQuestionHoverIcon.svg';
-import AiQuestion from '@svgs/component-set/AiQuestionIcon.svg';
-import ArrowLeft from '@svgs/component-set/ArrowLeftIcon.svg';
-import ArrowRight from '@svgs/component-set/ArrowRightIcon.svg';
-import CheckboxCircleDisable from '@svgs/component-set/CheckboxCircleDisableIcon.svg';
-import CheckboxCircleEnable from '@svgs/component-set/CheckboxCircleEnableIcon.svg';
-import CheckboxSquareDisable from '@svgs/component-set/CheckboxSquareDisableIcon.svg';
-import CheckboxSquareEnable from '@svgs/component-set/CheckboxSquareEnableIcon.svg';
-import Copy from '@svgs/component-set/CopyIcon.svg';
-import CrossLine from '@svgs/component-set/CrossLineIcon.svg';
-import DownloadDocs from '@svgs/component-set/DownloadDocsIcon.svg';
-import DownloadPdf from '@svgs/component-set/DownloadPdfIcon.svg';
-import DownloadWord from '@svgs/component-set/DownloadWordIcon.svg';
-import DownloadHangul from '@svgs/component-set/DownloadhangulIcon.svg';
-import Edit from '@svgs/component-set/EditIcon.svg';
-import EyeClose from '@svgs/component-set/EyeCloseIcon.svg';
-import EyeOpen from '@svgs/component-set/EyeOpenIcon.svg';
-import GnbAiManager from '@svgs/component-set/GnbAiManagerIcon.svg';
-import GnbCalender from '@svgs/component-set/GnbCalenderIcon.svg';
-import GnbHome from '@svgs/component-set/GnbHomeIcon.svg';
-import GnbLogout from '@svgs/component-set/GnbLogoutIcon.svg';
-import GnbProfile from '@svgs/component-set/GnbProfileIcon.svg';
-import GnbRecentFile from '@svgs/component-set/GnbRecentFileIcon.svg';
-import GnbSetting from '@svgs/component-set/GnbSettingIcon.svg';
-import GnbSnsAssist from '@svgs/component-set/GnbSnsAssistIcon.svg';
-import GnbTeamMood from '@svgs/component-set/GnbTeamMoodIcon.svg';
-import Link from '@svgs/component-set/LinkIcon.svg';
-import Plus from '@svgs/component-set/PlusIcon.svg';
-import RecordingEnd from '@svgs/component-set/RecordingEndIcon.svg';
-import RecordingPause from '@svgs/component-set/RecordingPauseIcon.svg';
-import RecordingPlay from '@svgs/component-set/RecordingPlayIcon.svg';
-import RecordingStart from '@svgs/component-set/RecordingStartIcon.svg';
-import SnsAssist from '@svgs/component-set/SnsAssistIcon.svg';
-import TeamMood from '@svgs/component-set/TeamMoodIcon.svg';
-import User1 from '@svgs/component-set/User1Icon.svg';
-import User2 from '@svgs/component-set/User2Icon.svg';
-import User3 from '@svgs/component-set/User3Icon.svg';
-import User4 from '@svgs/component-set/User4Icon.svg';
-import User5 from '@svgs/component-set/User5Icon.svg';
-import User6 from '@svgs/component-set/User6Icon.svg';
-import User7 from '@svgs/component-set/User7Icon.svg';
-import ProfileAdd from '@svgs/component-set/profileAdd.svg';
-import ProfileCircle from '@svgs/component-set/profileCircle.svg';
-import ArrowDown from '@svgs/individual/ArrowUnderIcon.svg';
-import Calender from '@svgs/individual/CalenderIcon.svg';
-import Congratulation from '@svgs/individual/CongratulationIcon.svg';
-import Download from '@svgs/individual/DownloadIcon.svg';
-import Search from '@svgs/individual/SearchIcon.svg';
-import Warning from '@svgs/individual/WarningIcon.svg';
-import LogoMixed from '@svgs/logo/LogoMixed.svg';
-import LogoSimbol from '@svgs/logo/LogoSimbol.svg';
-import LogoText from '@svgs/logo/LogoText.svg';
-
-const LogoTest = () => {
-  return (
-    <div className="grid grid-cols-10">
-      <AiManager height="100" width="100" />
-      <AiQuestionHover height="100" width="100" />
-      <AiQuestion height="100" width="100" />
-      <ArrowLeft height="100" width="100" />
-      <ArrowRight height="100" width="100" />
-      <CheckboxCircleDisable height="100" width="100" />
-      <CheckboxCircleEnable height="100" width="100" />
-      <CheckboxSquareDisable height="100" width="100" />
-      <CheckboxSquareEnable height="100" width="100" />
-      <Copy height="100" width="100" />
-      <CrossLine height="100" width="100" />
-      <DownloadDocs height="100" width="100" />
-      <DownloadPdf height="100" width="100" />
-      <DownloadWord height="100" width="100" />
-      <DownloadHangul height="100" width="100" />
-      <Edit height="100" width="100" />
-      <EyeClose height="100" width="100" />
-      <EyeOpen height="100" width="100" />
-      <GnbAiManager height="100" width="100" />
-      <GnbCalender height="100" width="100" />
-      <GnbHome height="100" width="100" />
-      <GnbLogout height="100" width="100" />
-      <GnbProfile height="100" width="100" />
-      <GnbRecentFile height="100" width="100" />
-      <GnbSetting height="100" width="100" />
-      <GnbSnsAssist height="100" width="100" />
-      <GnbTeamMood height="100" width="100" />
-      <ProfileAdd height="100" width="100" />
-      <ProfileCircle height="100" width="100" />
-      <Link height="100" width="100" />
-      <Plus height="100" width="100" />
-      <RecordingEnd height="100" width="100" />
-      <RecordingPause height="100" width="100" />
-      <RecordingPlay height="100" width="100" />
-      <RecordingStart height="100" width="100" />
-      <SnsAssist height="100" width="100" />
-      <TeamMood height="100" width="100" />
-      <User1 height="100" width="100" />
-      <User2 height="100" width="100" />
-      <User3 height="100" width="100" />
-      <User4 height="100" width="100" />
-      <User5 height="100" width="100" />
-      <User6 height="100" width="100" />
-      <User7 height="100" width="100" />
-      <ArrowDown height="100" width="100" />
-      <Calender height="100" width="100" />
-      <Congratulation height="100" width="100" />
-      <Download height="100" width="100" />
-      <Search height="100" width="100" />
-      <Warning height="100" width="100" />
-      <LogoMixed height="100" width="100" />
-      <LogoSimbol height="100" width="100" />
-      <LogoText height="100" width="100" />
-    </div>
-  );
-};
-
-export default LogoTest;
