+++ src/common/components/box-text-temp/SnsLink/SnsLink.client.tsx
