+++ src/common/components/box-text-temp/SnsLink/SnsLink.types.ts
