+++ src/common/components/box-text-temp/SurveyInfo/SurveyInfo.server.tsx
