+++ src/common/components/box-text-temp/SurveyInfo/SurveyInfo.types.ts
