+++ src/common/components/list-file-temp/ListFileSnsEventAssistant/ListFileSnsEventAssistant.client.tsx
