+++ src/common/components/list-file-temp/ListFileSnsEventAssistant/ListFileSnsEventAssistant.types.ts
