+++ src/common/components/list-file-temp/ListFileSnsEventAssistantLink/ListFileSnsEventAssistantLink.types.ts
