+++ src/common/components/list-file-temp/ListFileTeamMoodTracker/ListFileTeamMoodTracker.stories.tsx
