+++ src/common/components/survey-temp/BarChart/BarChart.stories.tsx
