+++ src/common/components/survey-temp/BarChart/BarChart.types.ts
