+++ src/common/components/survey-temp/BarChart/bar-chart-config.ts
