+++ src/common/components/survey-temp/BarChart/bar-chart-plugin.ts
