+++ src/common/components/survey-temp/SubjectiveAnswer/SubjectiveAnswser.server.tsx
@@ -1,4 +1,4 @@
-import ChartContainer from '@common/components/Survey/ChartContainer/ChartContainer.server';
+import ChartContainer from '@common/components/survey-temp/ChartContainer/ChartContainer.server';
 
 import { SubjectiveAnswersProps } from './SubjectiveAnswer.types';
